layer-update: fix musicpd, work around ntpd deps
This patch fixes a missing dependency of the meta-oe systemd layer to musicpd
by including meta-multimedia in the layer stack.
Another missing dependency of meta-systemd, ntp-systemd, is worked around:
the systemd layer's .bbappend refers to a deprecated (and by now nonexistent)
version of ntpd. To work around this missing dependency meta-hidav includes the
old recipe for now.