hidav-machine.sh: delete
Commit 4c3184563ab7d3e5fc7f902cf888fe84fac70eeb to build local.conf
from a template broke the script used to set the target machine,
hidav-machine.sh.

We have decided not to fix the script, but to permanently remove it.

Target machines will now be set manually in the config files.
The default machine should be set in local.conf.template and may
be overridden by assigning a value in private.conf.

Signed-off-by: Angelika Tobisch <tobisch@dresearch.de>