use PR server instead of decprecated PRINC, HYP-9203
This avoids warnings of this type:

"WARNING: Use of PRINC ... was detected in the recipe ... (or one
of its .bbappends) Use of PRINC is deprecated. The PR server should
be used to automatically increment the PR.
See: https://wiki.yoctoproject.org/wiki/PR_Service."

HINT: This patch results in a decrement of various package versions
which leads to an QA issue! We live with that because daisy branch
is not yet productive.

Signed-off-by: Steffen Sledz <sledz@dresearch-fe.de>